Yesdk-804 get assertion parameters (#505)
* start MakeCredentialParameters class

* add more Make Credential parameters

* fix merge issues

* fixed error in test

* update based on review

* fixed some issues

* add first draft of Make Credential Command

* update based on review

* more make credential response decoding

* add code to parse and process the MakeCredential response

* add GetAssertionParameters and update Cbor code

* update based on review

+++ Yubico.YubiKey/tests/unit/Yubico/YubiKey/Fido2/Cbor/CborHelperTests.cs
@@ -33,7 +33,7 @@ namespace Yubico.YubiKey.Fido2.Cbor
             byte[] encoded = w.Encode();
 
             var r = new CborReader(encoded, CborConformanceMode.Ctap2Canonical);
-            var m = new CborMap<long>(r);
+            var m = new CborMap<int>(r);
 
             Assert.Equal("test", m.ReadTextString(1));
             Assert.Equal("foo", m.ReadTextString(2));
